Fix silent UI and so on.
- I forgot to commit to git these changes which should be applied to #959.
