+++ OpenKh.Tools.ModsManager/Models/ViewHelper/ActionCommand.cs
@@ -7,7 +7,7 @@ using System.Windows.Input;
 
 namespace OpenKh.Tools.ModsManager.Models.ViewHelper
 {
-    public record ActionCommand(string Display, ICommand Command)
+    public record ActionCommand(string Display, string ToolTip, ICommand Command)
     {
     }
 }
