+++ OpenKh.Tools.ModsManager/Views/NotepadVM.cs
@@ -1,28 +1,38 @@
 using OpenKh.Tools.ModsManager.Services;
-using OpenKh.Tools.ModsManager.UserControls;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
 using System.Windows.Input;
 using Xe.Tools;
-using Xe.Tools.Wpf.Commands;
-using Xe.Tools.Wpf.Dialogs;
 
 namespace OpenKh.Tools.ModsManager.Views
 {
     public class NotepadVM : BaseNotifyPropertyChanged
     {
         public ColorThemeService ColorTheme => ColorThemeService.Instance;
-        public ICommand CopyAllCommand { get; set; }
-        public ICommand SaveAsCommand { get; set; }
 
-        private readonly GetActiveWindowService _getActiveWindowService = new GetActiveWindowService();
+        #region CopyAllCommand
+        private ICommand _copyAllCommand = null;
+        public ICommand CopyAllCommand
+        {
+            get => _copyAllCommand;
+            set
+            {
+                _copyAllCommand = value;
+                OnPropertyChanged(nameof(CopyAllCommand));
+            }
+        }
+        #endregion
 
-        private string SavedTo { get; set; }
+        #region SaveAsCommand
+        private ICommand _saveAsCommand = null;
+        public ICommand SaveAsCommand
+        {
+            get => _saveAsCommand;
+            set
+            {
+                _saveAsCommand = value;
+                OnPropertyChanged(nameof(SaveAsCommand));
+            }
+        }
+        #endregion
 
         #region Text
         private string _text = "";
@@ -36,47 +46,5 @@ namespace OpenKh.Tools.ModsManager.Views
             }
         }
         #endregion
-
-        public NotepadVM()
-        {
-            CopyAllCommand = new RelayCommand(
-                _ =>
-                {
-                    try
-                    {
-                        Clipboard.SetText(Text);
-                    }
-                    catch (Exception ex)
-                    {
-                        MessageBox.Show("Failed to copy!\n\n" + ex);
-                    }
-                }
-            );
-
-            SaveAsCommand = new RelayCommand(
-                _ =>
-                {
-                    FileDialog.OnSave(
-                        path =>
-                        {
-                            SavedTo = path;
-
-                            try
-                            {
-                                File.WriteAllText(path, Text);
-                            }
-                            catch (Exception ex)
-                            {
-                                MessageBox.Show("Failed to save to file!\n\n" + ex);
-                            }
-                        },
-                        new List<FileDialogFilter>()
-                            .AddAllFiles(),
-                        SavedTo,
-                        _getActiveWindowService.GetActiveWindow()
-                    );
-                }
-            );
-        }
     }
 }
